fix(RVCV-147): JPQL로 수정 및 쿼리 변경
